feat(abi)!: widen language codes to u32, format version 3
BREAKING: .fzm format 2 artifacts no longer load, and fizh_translate /
fizh_can_translate take u32 language codes. Re-run tools/fetch-model.sh; there
is nothing to migrate (docs/ABI.md).

The u16 packed-ASCII-pair convention was this project's invention and the
registry does not honour it. ISO 639-3 codes are three letters, and Firefox
ships script-qualified pairs -- zh-Hans, zh-Hant -- that two bytes cannot
express at all. zh-Hans-en was refused for exactly that reason and nothing
else.

abi.Lang is now u32: one to four lowercase ASCII bytes, big-endian, zero-padded
on the left, with a validator that rejects a zero wedged between two letters
because padding is on the left or it is corruption. Four bytes still cannot
hold "zh-Hans" verbatim, so the converter maps registry codes that are not
639-1 to a fizh short form -- zh-Hans -> zhs, zh-Hant -> zht (convert.LANG_SHORT).
The code is an identifier for routing, not a BCP-47 tag; it has to be unique
and stable, not expressive.

The header grows 64 -> 128 bytes to fit two u32 codes, with 60 bytes reserved
so the next field does not cost another version. Four readers of that header
existed -- format.zig, convert.py, reference.py, and two JS hosts -- which is
itself worth noting.

zh-Hans-en now translates: "The black cat was sleeping on the table."
It is 48.19 MB, over even the raised 35 MB budget, which is a separate fact.

+++ src/root.zig
@@ -47,8 +47,8 @@ export fn fizh_translate(
     h: i32,
     src: [*]const u8,
     src_len: u32,
-    src_lang: u16,
-    tgt_lang: u16,
+    src_lang: u32,
+    tgt_lang: u32,
     out: [*]u8,
     out_cap: u32,
 ) i32 {
@@ -59,7 +59,7 @@ export fn fizh_translate(
 
 /// 0 = no, 1 = direct, 2 = pivot. Lets the host grey out a language before the
 /// user picks it.
-export fn fizh_can_translate(h: i32, src_lang: u16, tgt_lang: u16) i32 {
+export fn fizh_can_translate(h: i32, src_lang: u32, tgt_lang: u32) i32 {
     return runtime.canTranslate(h, src_lang, tgt_lang);
 }
